Refactor Bencode into Reader/Writer pair mirroring TOML
Split the monolithic static `Bencode` class into a thin facade delegating
to new `BencodeReader` (with `BencodeReader.Parser.cs`) and `BencodeWriter`
types, matching the `Toml`/`TomlReader`/`TomlWriter` architecture.

- `BencodeReader` owns decoding: `Read`/`TryRead` span overloads (threading
  `BencodeParseOptions`), plus `byte[]`/`Stream`/async stream overloads and
  the moved `Parser` ref struct.
- `BencodeWriter` owns encoding: `Write`/`TryWrite`/`GetWrittenLength` plus
  `Stream`/async stream overloads and the encode helpers.
- `Bencode` retains every existing public signature as one-line delegations,
  so existing tests and `BencodeExtensions` are unaffected.
- Removed `Bencode.Stream.cs`; its overloads moved onto the reader/writer.

Value types stay immutable; `BencodeParseOptions` and `BencodeExtensions`
are kept. Added `BencodeReaderWriterTests` covering the new surface.
